storage with committed version list

+++ common/net.hpp
@@ -20,7 +20,7 @@ namespace net
     std::string receive(StreamSocket& socket)
     {
         unsigned char buffer[sizeof(size_t)];
-        socket.receiveBytes(buffer, sizeof(buffer));
+        socket.receiveBytes(buffer, (int)sizeof(buffer));
         size_t size = *(size_t*)&buffer;
 
         std::string msg;
@@ -28,7 +28,7 @@ namespace net
         int received_bytes = 0;
         while (received_bytes < size)
         {
-            received_bytes += socket.receiveBytes(msg.data() + received_bytes, size - received_bytes);
+            received_bytes += socket.receiveBytes(msg.data() + received_bytes, (int)(size - received_bytes));
         }
         return msg;
     }
